add vim config from cadence vm vim to mac
note that the plugins in the bundle directory are empty - this is
expected behavior as each one is its own git repository. to load them on
to a new setup, clone the modules from github. this can be automated if
we use vundle, but we're using pathogen for its simplicity at the
moment.